This morning the module only worked for messages <512 bits. I learned a few things troubleshooting this. First, the input message always gets a zero pad, even if the message is a perfect 512 bits. I thought it wouldn't be necessary, but careful reading of the standard never mentions that the pad is optional. This wasn't a difficult fix once I figured it out. Second, I had a very subtle issue that was causing multi-block messages to fail. Before going to compress the next block, the working variables are set to the value of the last hash. One of the working variables was getting hash_vals[1] when it should have been getting hash_vals[2]. I had to look closely at the code to see it, because there are a bunch of assignments happening before iterating to the next compression. Attention to detail is key!
I also tested the module with messages of the following lengths (in words): 3, 10, 14, 15, 16, 35, 100. I figured going past 100 really wasn't going to give me much. Note, it's required for this module that the message be a round number of words. It's easy to mess up either the input to a SHA-256 calculator or the test in test_functions.c and get (what appears to be) a failing result. I consider this module done for the time being. I may come back and improve it later.